Change node Float to Double

+++ src/osmproc/Node.java
@@ -2,8 +2,8 @@ package osmproc;
 
 public class Node {
     private String id;
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
 
     public String getId() {
         return id;
@@ -13,19 +13,19 @@ public class Node {
         this.id = id;
     }
 
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
